Refactor

+++ craps-simulator/Models/BetResult.cs
@@ -6,7 +6,7 @@ namespace craps_simulator.Models {
         public decimal Bet { get; set; } = 0;
         public bool IsWinner { get; set; } = false;
         public bool IsLoser { get; set; } = false;
-        public int Pays { get; set; } = 0;
+        public decimal Pays { get; set; } = 0;
         public string Msg { get; set; } = string.Empty;
     }
 }
